Move resetting of company details to a concern for reuse
