+++ app/controllers/waste_carriers_engine/business_type_forms_controller.rb
@@ -2,6 +2,8 @@
 
 module WasteCarriersEngine
   class BusinessTypeFormsController < ::WasteCarriersEngine::FormsController
+    include CanResetCompanyDetails
+
     def new
       super(BusinessTypeForm, "business_type_form")
     end
@@ -15,15 +17,6 @@ module WasteCarriersEngine
 
     def transient_registration_attributes
       params.fetch(:business_type_form, {}).permit(:business_type, :token)
-    end
-
-    # Clear any previous business-type specific attributes to handle cases where the user
-    # starts with one business type and then navigates back and changes the business type.
-    def reset_company_attributes
-      @transient_registration.company_no = nil
-      @transient_registration.registered_company_name = nil
-      @transient_registration.temp_use_registered_company_details = nil
-      @transient_registration.save!
     end
   end
 end
